// src/app/shared/interfaces/icourtowner-courts.ts
export interface IcourtownerCourts {
    id: number
    maincourt_id: number
    name: string
    description: string
    type: string
    surface_type: string
    price_per_hour: string
    status: string
    is_open: boolean
    created_at: string
    updated_at: string
    primary_image: any
    timeslots: any[]
}